// src/shared/enums/OrderStatus.enum.ts
export enum OrderStatusEnum {
  STATUS_BOOKED = 'STATUS_BOOKED',
  STATUS_STARTED = 'STATUS_STARTED',
  STATUS_ARRIVED = 'STATUS_ARRIVED',
  STATUS_CHECKED = 'STATUS_CHECKED',
  STATUS_LOADING = 'STATUS_LOADING',
  STATUS_LOADED = 'STATUS_LOADED',
  STATUS_PICKED_UP = 'STATUS_PICKED_UP',
  STATUS_REACHED = 'STATUS_REACHED',
  STATUS_UNLOADING = 'STATUS_UNLOADING',
  STATUS_UNLOADED = 'STATUS_UNLOADED',
  STATUS_COMPLETE = 'STATUS_COMPLETE',
  STATUS_CLIENT_PAID = 'STATUS_CLIENT_PAID',
  STATUS_PAID = 'STATUS_PAID',
  STATUS_CANCELED = 'STATUS_CANCELED',
  STATUS_ASSIGNED = 'STATUS_ASSIGNED',
}